requirements.txt
streamlit==1.31.0
mlx>=0.18.0
mlx-lm>=0.18.0
click>=8.1.0
huggingface_hub>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
regex>=2023.0.0
watchdog>=4.0.0
plotly>=5.17.0
requests>=2.31.0
transformers>=4.36.0
tokenizers>=0.15.0
datasets>=2.16.0
psutil>=5.9.0
tqdm>=4.66.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.42.0
scikit-learn>=1.3.0